Fix PSalm
