+++ tests/Fixture/MyUserModel.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Tests\Fixture;
+
+use ByJG\Authenticate\Model\UserModel;
+
+class MyUserModel extends UserModel
+{
+    protected $otherfield;
+
+    public function __construct($name = "", $email = "", $username = "", $password = "", $admin = "no", $field = "")
+    {
+        parent::__construct($name, $email, $username, $password, $admin);
+        $this->setOtherfield($field);
+    }
+
+    public function getOtherfield()
+    {
+        return $this->otherfield;
+    }
+
+    public function setOtherfield($otherfield): void
+    {
+        $this->otherfield = $otherfield;
+    }
+}
